requirements.txt
google-generativeai
PyMuPDF
gradio
